% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drone_image_area.R
\name{drone_image_area}
\alias{drone_image_area}
\title{Calculate the area of land captured in a drone image}
\usage{
drone_image_area(
  camera_angle,
  vertical_fov,
  horizontal_fov,
  height,
  pixel_row,
  pixel_col
)
}
\arguments{
\item{camera_angle}{Angle of the camera in degrees, as measured from the
height of the drone.}

\item{vertical_fov}{Vertical field of view of the camera, in degrees.}

\item{horizontal_fov}{Horizontal field of view of the camera, in degrees.}

\item{height}{Height of the drone in metres.}

\item{pixel_row}{Number of rows of pixels recorded.}

\item{pixel_col}{Number of columns of pixels recorded.}
}
\value{
Area of the trapezoid captured in the image, measured in metres
squared.
}
\description{
Calculate the area of land captured in a drone image
}
